requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
alpha_vantage
streamlit-chat
python-dotenv